+++ CHANGELOG.md
@@ -22,7 +22,8 @@ All user visible changes to this project will be documented in this file. This p
 - Signalling:
     - Dynamic `Peer`s creation when client connects ([#28]);
     - Auto-removing `Peer`s when `Member` disconnects ([#28]);
-    - Filter `SetIceCandidate` messages without `candidate` ([#50](/../../pull/50)).
+    - Filter `SetIceCandidate` messages without `candidate` ([#50](/../../pull/50));
+    - Send reason of WebSocket close by server as close frame's description.
 - Testing:
     - E2E tests for signalling ([#28]).